C UD 공통 응답 클래스 ApiResponse 생성, 서비스 로직 일부 수정

+++ back/src/main/java/PHCCS/service/post/PostService.java
@@ -190,6 +190,10 @@ public class PostService {
     public void deletePost(String category, Long memberId, Long postId){
         log.info("|se|deletePost()");
         String findFileDir = repository.findPostDir(category, postId);
+
+        Post post = repository.showPost(category, postId);
+        if(post == null) throw new BadRequestEx("게시글이 존재하지 않습니다.");
+
         repository.deletePost(category, memberId, postId);
 
         if(findFileDir != null){
